Remove multtable, replace with logical equivalent
Multtable stored the result of a 8-bit signed multiply of the
upper nibble of A (interpreted as signed) and the amplitude
lower nibble (interpreted as unsigned), then divided by two.
On the 6510 this made sense, but in modern processors it's
way faster and cleaner to simply do the multiply.

Replace the render sin/sin/rect mixer with equivalent C code.